Attempted refactoring of random failed
Attempted refactoring of random failed yet again.
This time I had a kernel launch time error. Not sure
what takes kernel to launch so long in the absence
of curand states but I am not going to touch it for know
I had already advanced not bad with removal of free_world
kernel.

Now I should finish up that triangle and cornell box