Passing groups to hittables work
This commit marks the solution to my design issues
with respect to hittable groups and their effects.

Basically the solution was hinted at nvidias post
about acceleration structures.
In their post they talked about lower level
acceleration structure and a higher level
acceleration structure.

At first I did not quite understand why we need
to divide the code base into these two levels.
But now it is more or less clear.

Low level structure holds all the building blocks
of the scene, it is a huge blob of surfaces.

High level structure holds all the information
about objects that are made by those surfaces.
It is this high level structure that we iterate.
This high level structure also organizes the
scattering. It is not very intuitive
because we think of scatterin in terms of solids.
But for liquids and gas, the volume in which
the material subsides also effects how the
scattering would take place.
Meaning that at least in our design scattering
involves multiple surfaces, though physically
this is not the case.