DBSCAN almost done requires hittable construction.
Dbscan acceleration structure is almost done.
It only needs to construct hittables using
labels. The methods provide the facilities for that.
The idea is we find the size of the label.
Initialize an array for it than add Hittable*
to it. These groups should be what ? Hittables,
HittableGroup, Hittable*, I am not so sure now.
Hittables is the most reasonable in terms of semantics
HittableGroup is just a wrapper over the Hittable*
blob. DBSCAN is a top level acceleration structure.
Thus essentially it would be working with
HittableGroup instead of triangles when initialized.