Refac almost done. Requires grouping logic
Refac almost done. The only thing that is left is to
implement some sort of grouping logic.
The logic should be something like the following.
- GroupType | INSTANCE | CONSTANT_MEDIUM
- group_id :int

from this we organise scene primitives into groups.
Then when the scene objects output a list of hittables,
we output groups.
As an acceleration structure, we are going to use
dbscan, which means we change the groups of surfaces.
For that we need to add a new dbscan id to scene objects.
The general idea is that we group primitives into objects,
which can have a group type of INSTANCe or CONSTANT_MEDIUM.
Then we group those objects into chunks. Each chunk is
attributed to a set by dbscan.
The to_hittable_list of SceneObjects output these
chunks.